+++ GMlib/modules/scene/src/sceneobjects/gmsphere3d.h
@@ -285,17 +285,17 @@ namespace GMlib {
       *ptr_n++ = top.getDir()(1);
       *ptr_n++ = top.getDir()(2);
 
-      u = M_PI_2 - du;
+      u = M_PI_2 - double(du);
 
-      su = sin(u);
-      cu = cos(u);
+      su = sin(double(u));
+      cu = cos(double(u));
       ru = r * cu;
 
       float v = 0;
       for( int i = 0; i < _m2+1; i++, v += dv ) {
 
-        const float sv = sin(v);
-        const float cv = cos(v);
+        const float sv = sin(double(v));
+        const float cv = cos(double(v));
 
         *ptr_v++ = x + ru*cv;
         *ptr_v++ = y + ru*sv;
@@ -314,16 +314,16 @@ namespace GMlib {
       *ptr_n++ = bottom.getDir()(1);
       *ptr_n++ = bottom.getDir()(2);
 
-      u = M_PI_2 - du * (_m1-1);
-      su = sin(u);
-      cu = cos(u);
+      u = M_PI_2 - double(du) * (_m1-1);
+      su = sin(double(u));
+      cu = cos(double(u));
       ru = r * cu;
 
       v = 0;
       for( int i = _m2; i >= 0; i--, v += dv ) {
 
-        const float sv = sin(v);
-        const float cv = cos(v);
+        const float sv = sin(double(v));
+        const float cv = cos(double(v));
 
         *ptr_v++ = x + ru*cv;
         *ptr_v++ = y + ru*sv;
@@ -338,22 +338,22 @@ namespace GMlib {
     // Compute body triangle strips on the sphere
     for( int i = 0; i < _m1-2; i++ ) {
 
-      const float u1 = M_PI_2 - du * (i+1);
-      const float u2 = M_PI_2 - du * (i+2);
+      const float u1 = M_PI_2 - double(du) * (i+1);
+      const float u2 = M_PI_2 - double(du) * (i+2);
 
-      const float su1 = sin(u1);
-      const float cu1 = cos(u1);
+      const float su1 = sin(double(u1));
+      const float cu1 = cos(double(u1));
       const float ru1 = r * cu1;
 
-      const float su2 = sin(u2);
-      const float cu2 = cos(u2);
+      const float su2 = sin(double(u2));
+      const float cu2 = cos(double(u2));
       const float ru2 = r * cu2;
 
       for( int j = 0; j < _m2+1; j++ ) {
 
         const float v = j * dv;
-        const float sv = sin(v);
-        const float cv = cos(v);
+        const float sv = sin(double(v));
+        const float cv = cos(double(v));
 
         *ptr_v++ = x + ru1*cv;
         *ptr_v++ = y + ru1*sv;
